完成了图片文件处理逻辑，增加临时文件夹概念 完成商品排序功能

+++ src/admin/controller/goods.js
@@ -1,6 +1,7 @@
 const Base = require('./base.js');
 const Config = require('../../common/config/config');
 const Upload = require('./upload.js');
+const FileUtil = require('../../common/fileutil');
 
 module.exports = class extends Base {
   /**
@@ -9,11 +10,19 @@ module.exports = class extends Base {
    */
   async indexAction() {
     const page = this.get('page') || 1;
-    const size = this.get('size') || 10;
+    const size = this.get('size') || 20;
     const name = this.get('name') || '';
+    const category_id = this.get('category_id') || '';
 
     const model = this.model('goods');
-    const data = await model.where({name: ['like', `%${name}%`]}).order(['id DESC']).page(page, size).countSelect();
+
+    const params = {name: ['like', `%${name}%`]};
+    if(!think.isEmpty(category_id))
+    {
+      params.category_id=category_id;
+    }
+    
+    const data = await model.where(params).order(['sort_order ASC']).page(page, size).countSelect();
 
     return this.success(data);
   }
@@ -51,20 +60,28 @@ module.exports = class extends Base {
     }
 
     const values = this.post();
-    const id = this.post('id');
+    let id = this.post('id');
 
     const model = this.model('goods');
     values.is_on_sale = values.is_on_sale ? 1 : 0;
     values.is_new = values.is_new ? 1 : 0;
     values.is_hot = values.is_hot ? 1 : 0;
 
-    values.goods_desc = JSON.stringify(values.goods_desc);
+    values.goods_desc = FileUtil.moveTmpImgToFinal(values.goods_desc); // 将商品详情图片移动到正式目录
 
+    let movedPosterImgs = FileUtil.moveTmpImgToFinal(values.list_pic_url); // 将商品封面移动到正式目录
+    if(movedPosterImgs && movedPosterImgs.length>0)
+    {
+      values.list_pic_url = movedPosterImgs[0];
+    }
+
+    values.goods_desc = JSON.stringify(values.goods_desc);
+    
     if (id > 0) {
       await model.where({id: id}).update(values);
     } else {
       delete values.id;
-      await model.add(values);
+      id = await model.add(values);
     }
 
     // 更新商品banner
@@ -72,8 +89,17 @@ module.exports = class extends Base {
     if(gallery)
     {
       gallery.map((item)=>{
+
+
         let galleryModel = this.model('goods_gallery');
         item.goods_id = id;
+
+        let movedImgs = FileUtil.moveTmpImgToFinal(item.img_url); // 将商品banner移动到正式目录
+        if(movedImgs && movedImgs.length>0)
+        {
+          item.img_url = movedImgs[0];
+        }
+
         if(item.id >0 )
         {
           galleryModel.where({goods_id: item.goods_id}).update(item);
@@ -82,30 +108,65 @@ module.exports = class extends Base {
         }
       });
     }
+
     
     if(values.deletedGalleries && values.deletedGalleries.length>0)
     {
       let deletedGalleryPics = [];
       values.deletedGalleries.map((item)=>{
+        
+        if(item.id>0)
+        {
+          this.model('goods_gallery').where({id: item.id}).limit(1).delete();
+        }
         deletedGalleryPics.push(item.img_url);
       });
 
         // 删除服务器中的商品banner图片文件
-        Upload.deleteImg(deletedGalleryPics);
+        FileUtil.deleteImg(deletedGalleryPics);
     }
   
 
     // 删除服务器中的详情图片文件
-    Upload.deleteImg(values.deletedDescPics);
+    FileUtil.deleteImg(values.deletedDescPics);
 
     return this.success(values);
   }
 
   async destoryAction() {
     const id = this.post('id');
-    await this.model('goods').where({id: id}).limit(1).delete();
+    let obj = await this.model('goods').where({id: id}).limit(1).find();
     // TODO 删除图片
 
+    if(obj)
+    {
+
+      // 删除封面
+      FileUtil.deleteImg(obj.list_pic_url);
+
+      // 删除详情
+      try{
+        obj.goods_desc = JSON.parse(obj.goods_desc);
+      }catch(e)
+      {
+        obj.goods_desc = [];
+      }
+      FileUtil.deleteImg(obj.goods_desc);
+
+      // 删除banner
+      const gallery = await this.model('goods_gallery').where({goods_id: id}).select();
+
+      if(gallery)
+      {
+        gallery.map((item)=>{
+          FileUtil.deleteImg(item.img_url);
+        });
+
+        this.model('goods_gallery').where({goods_id: id}).delete(); // 删除banner数据
+      }
+
+      this.model('goods').where({id: id}).limit(1).delete(); // 删除商品数据
+    }
     return this.success();
   }
 };
